Added option to use absolute path to LDS data filed. JIRA: CXX-1322
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@45194 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/objtools/lds/lds_object.cpp
@@ -228,11 +228,13 @@ void CLDS_Object::DeleteCascadeFiles(const CLDS_Set& file_ids,
     CBDB_FileCursor cur(m_db.annot_db);
     cur.SetCondition(CBDB_FileCursor::eFirst);
     while (cur.Fetch() == eBDB_Ok) {
-        int fid = m_db.object_db.file_id;
-        if (fid && LDS_SetTest(file_ids, fid)) {
-            int annot_id = m_db.annot_db.annot_id;
-            annotations_deleted->set(annot_id);
-            m_db.annot_db.Delete();
+        if ( !m_db.object_db.file_id.IsNull() ) {
+            int fid = m_db.object_db.file_id;
+            if (fid && LDS_SetTest(file_ids, fid)) {
+                int annot_id = m_db.annot_db.annot_id;
+                annotations_deleted->set(annot_id);
+                m_db.annot_db.Delete();
+            }
         }
     }
